(Mostly) fix the frequent glitches in the picture stability.
Seems the hall sensor was getting stray signals from the motor coils or the
cable supplying current to the motor.

Attempt to filter that out by ignoring periods that seem to be too far off
what we expect.

I still saw a couple of glitches shortly after start, so might need some
fine-tuning, but it seems to help a lot, at least.