Adjust colour contrast, _much_ better now.
Space apart the higher colour numbers more, and the lower less, to
make the perceived difference between adjacent colour numbers 0-15
more equal. This greatly improves the contrast of eg. the displayed
cartoons.

Also adjust the angle to make the picture orientation correct and
attempt to tune the de-glitch values.

There is still an occasional glitch, but it's surely much improved
now, especially with the contrast fix.