Sync the TLC latching interrupt to the hall signal to improve picture stability.
In general, the rotation period will not be an integer multiple of the
PWM/latching period. But we want each pixel to ideally turn on and off at the
same spatial position in each refresh.

So implement synchonisation to achieve this (assuming stable rotation and hall
signal). When we get a hall signal, adjust the coming latching interrupt
period to make latching trigger at an integer number of PWM periods after the
time of the hall signal.

Also tweak the hall de-glitch parameters a bit, seems to help but more testing
will be needed to know for sure.