Implement reading animations off micro-SD card.
When animation data is received over nRF24L01+ wireless, play that.
When wireless is idle for some time, switch to playing animations read
from SD-card, if present; else play on-board animations.

The SD-card reading uses an interrupt for the bulk block loading, to
not get stalled by scanline-interrupt (the TM4C123x has an errata that
prevents using DMA on all the SSIs at the same time).